Update Get-CsOnlineVoiceUser.md
This documentation is misleading as it suggests the command will still work when it does not. 

The term "deprecated" usually serves as a warning that you can still use the method but there's a better way to do it. In this case, the command  `Get-CsOnlineVoiceUser` does not work at all anymore.